+++ datalogistik/__init__.py
@@ -13,3 +13,7 @@
 # limitations under the License.
 
 __version__ = "0.1.0"
+
+from .dataset import Dataset
+
+__all__ = ["Dataset"]
