+++ datalogistik/datalogistik.py
@@ -57,112 +57,9 @@ def main():
         finish()
 
     # At this point, we don't have any matches, so we need to bail
-
-    # local_cache_location = config.get_cache_location()
-    # log.debug(f"Checking local cache at {local_cache_location}")
-    # cached_dataset_path = util.create_cached_dataset_path(
-    #     argument_info.dataset,
-    #     argument_info.scale_factor,
-    #     argument_info.format,
-    #     str(argument_info.partition_max_rows),
-    #     argument_info.compression,
-    # )
-    # cached_dataset_metadata_file = pathlib.Path(
-    #     cached_dataset_path, config.metadata_filename
-    # )
-    # if cached_dataset_metadata_file.exists():
-    #     log.debug(f"Found cached dataset at '{cached_dataset_metadata_file}'")
-    #     util.output_result(cached_dataset_path)
-    #     finish()
-
-    # # Do not convert a tpc dataset to CSV, because Arrow cannot cast decimals to string (ARROW-17458)
-    # elif not (
-    #     argument_info.dataset in tpc_info.tpc_datasets and argument_info.format == "csv"
-    # ):
-    #     # not found in cache, check if the cache has other formats of this dataset
-    #     for cached_file_format in [x for x in config.supported_formats]:
-    #         if argument_info.scale_factor:
-    #             scale_factor = f"scalefactor_{argument_info.scale_factor}"
-    #         else:
-    #             scale_factor = ""
-
-    #         other_format_path = pathlib.Path(
-    #             local_cache_location,
-    #             argument_info.dataset,
-    #             scale_factor,
-    #             cached_file_format,
-    #         )
-    #         if other_format_path.exists():
-    #             log.debug(
-    #                 "Found cached instance(s) with a different format/partitioning/compression at "
-    #                 f"'{other_format_path}'"
-    #             )
-    #             format_subfolders = [
-    #                 f.name for f in os.scandir(other_format_path) if f.is_dir()
-    #             ]
-    #             # Grab the first subdir, we're going to convert it anyway
-    #             cached_nrows_dir = format_subfolders[0]
-    #             similar_dataset_path = other_format_path / cached_nrows_dir
-    #             cached_nrows = int(cached_nrows_dir.split("_")[-1])
-    #             nrows_subfolders = [
-    #                 f.name for f in os.scandir(similar_dataset_path) if f.is_dir()
-    #             ]
-    #             cached_compression_dir = nrows_subfolders[0]
-    #             similar_dataset_path = similar_dataset_path / cached_compression_dir
-    #             cached_compression = cached_compression_dir.split("_")[-1]
-
-    #             cached_dataset_metadata_file = pathlib.Path(
-    #                 similar_dataset_path, config.metadata_filename
-    #             )
-    #             if cached_dataset_metadata_file.exists():
-    #                 log.debug(
-    #                     "Found cached dataset in different format/partitioning/compression at "
-    #                     f"'{similar_dataset_path}'"
-    #                 )
-    #                 # cached_dataset_path = util.convert_dataset(
-    #                 #     dataset,
-    #                 #     cached_compression,
-    #                 #     argument_info.compression,
-    #                 #     cached_file_format,
-    #                 #     argument_info.format,
-    #                 #     cached_nrows,
-    #                 #     argument_info.partition_max_rows,
-    #                 # )
-    #                 util.output_result(cached_dataset_path)
-    #                 finish()
-    #             else:
-    #                 log.info("Found cached dataset without metadata file, cleaning...")
-    #                 util.clean_cache_dir(similar_dataset_path)
-
-    # # If we have not exited at this point, nothing useable was found in the local cache.
-    # # We need to either generate or download the data.
-    # log.info("Dataset not found in local cache.")
-
-    # if argument_info.dataset in tpc_info.tpc_datasets:
-    #     cached_dataset_path = util.generate_dataset(dataset, argument_info)
-    # else:
-    #     cached_dataset_path = util.download_dataset(dataset)
-
-    # if dataset["format"] == "parquet":
-    #     # retrieve the compression from the directory name
-    #     compression = cached_dataset_path.parts[-1].split("_")[-1].lower()
-    # else:
-    #     compression = dataset.get("file-compression")
-    # # Convert if necessary
-    # if (dataset["format"] != argument_info.format) or (
-    #     dataset["partitioning-nrows"] != argument_info.partition_max_rows
-    #     or (compression != argument_info.compression)
-    # ):
-    #     pass
-    #     # cached_dataset_path = util.convert_dataset(
-    #     #     dataset,
-    #     #     compression,
-    #     #     argument_info.compression,
-    #     #     dataset["format"],
-    #     #     argument_info.format,
-    #     #     dataset["partitioning-nrows"],
-    #     #     argument_info.partition_max_rows,
-    #     # )
+    log.debug(
+        f"Could not find any datasets that are feasible for {dataset.name}. Please check the name."
+    )
 
 
 if __name__ == "__main__":
